Use correct architectures in library.properties
The library uses AVR-specific code and will not compile for any other architecture. Therefore, the previous use of a wildcard architectures value was incorrect.

Use of a correct architectures value will cause the Arduino IDE to place the library's examples under File > Examples > INCOMPATIBLE. It will also give a helpful warning on compilation of the library:

WARNING: library Grove_Air_quality_Sensor-master claims to run on (avr) architecture(s) and may be incompatible with your current board which runs on (samd) architecture(s).

These will give the user valuable clues as to the cause of the somewhat obscure compilation errors.